Bug found in TimeSheets render() method.
Because it renders multiple times, the same operation is being
executed multiple times, adding that many times the duration when
it shouldn't.

Solution: run all the operation from the root data (this.state.db)
and store each step in a new member, so that when the project
filter is updated, the change affects in cascade to the following
steps, but doesn't re-operate in the changed state (increasing
the effect of the change!).